requirements.txt
requests
colorama
python-dotenv
pytelegrambotapi